+++ automerge-backend/src/op_handle.rs
@@ -4,8 +4,8 @@ use std::hash::{Hash, Hasher};
 use std::ops::Deref;
 use std::rc::Rc;
 
-use crate::protocol::{Change, Key, ObjectID, OpType, Operation, UndoOperation, Value};
-use automerge_protocol::OpID;
+use crate::protocol::{Change, Key, OpType, Operation, UndoOperation, Value};
+use automerge_protocol::{OpID, ObjectID};
 
 #[derive(Clone)]
 pub(crate) struct OpHandle {
