+++ automerge-protocol/src/serde_impls/mod.rs
@@ -1,2 +1,3 @@
 mod change_hash;
 mod opid;
+mod object_id;
