+++ automerge-protocol/src/utility_impls/mod.rs
@@ -1,1 +1,2 @@
 mod opid;
+mod object_id;
